refactor: merge card into one component - closes #125
